pyproject.toml
[project]
name = "astrazeneca"
version = "0.1.0"
description = ""
authors = [
    {name = "juan marcos",email = "juan.mervi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
biopython = ">=1.85,<2.0"
requests = "^2.32.3"
biotite = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
